Crate iroh_tickets

Tickets is a serializable object combining information required for an operation. Typically tickets contain all information required for an operation, e.g. an iroh blob ticket would contain the hash of the data as well as information about how to reach the provider.

Modules

endpoint

Tickets for endpoints.

Enums

ParseError

An error deserializing an iroh ticket.

Traits

Ticket

A ticket is a serializable object combining information required for an operation.